requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
kafka-python==2.0.2
lxml @ {path to your file}
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
